backend/pyproject.toml
[project]
name = "rss-backend"
version = "0.1.0"
description = "Local FastAPI service for RSS READER (feeds, storage, AI proxy)."
authors = [{ name = "RSS READER" }]
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlmodel>=0.0.21",
  "alembic>=1.13",
  "pydantic-settings>=2.7",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "feedparser>=6.0",
  "beautifulsoup4>=4.12",
  "readability-lxml>=0.8",
  "lxml>=5.3",
  "apscheduler>=3.10",
  "loguru>=0.7",
  "eval-type-backport>=0.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.6",
  "mypy>=1.11"
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
